pyproject.toml
[tool.poetry]
name = "details_shop_me"
version = "0.1.0"
description = ""
authors = ["givinity <SaZh.user@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
Pillow = "^9.5.0"
flake8 = "^6.0.0"
python-dotenv = "^1.0.0"
requests = "^2.30.0"
urllib3 = "<=2.0"
psycopg = "^3.1.9"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
